+++ plugins/main/public/utils/applications.ts
@@ -687,7 +687,7 @@ export const Applications = [
   googleCloud,
   office365,
   github,
-  // docker,
+  docker,
   endpointSummary,
   endpointGroups,
   serverStatus,